Don't group styled-components in one file
